pyproject.toml
[project]
name = "doomdiffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "click>=8.2.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "omgifol>=0.5.1",
    "pillow>=11.3.0",
    "ruff>=0.12.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
